requirements.txt
Flask>=2.0
oracledb>=1.0
python-dotenv